requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.109.2
google-ai-generativelanguage==0.4.0
google-api-core==2.24.0
google-auth==2.37.0
google-generativeai==0.3.2
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.62.3
h11==0.14.0
idna==3.10
motor==3.6.0
passlib==1.7.4
proto-plus==1.25.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pymongo==4.9.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
requests==2.31.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.36.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.27.1
